[FLINK-20640][python] Use cython to optimize the performance of Python UDAF
This closes #14449.
