+++ flink-python/setup.py
@@ -66,6 +66,10 @@ else:
                 sources=["pyflink/fn_execution/coder_impl_fast.pyx"],
                 include_dirs=["pyflink/fn_execution/"]),
             Extension(
+                name="pyflink.fn_execution.aggregate_fast",
+                sources=["pyflink/fn_execution/aggregate_fast.pyx"],
+                include_dirs=["pyflink/fn_execution/"]),
+            Extension(
                 name="pyflink.fn_execution.stream",
                 sources=["pyflink/fn_execution/stream.pyx"],
                 include_dirs=["pyflink/fn_execution/"]),
@@ -88,6 +92,10 @@ else:
                 Extension(
                     name="pyflink.fn_execution.coder_impl_fast",
                     sources=["pyflink/fn_execution/coder_impl_fast.c"],
+                    include_dirs=["pyflink/fn_execution/"]),
+                Extension(
+                    name="pyflink.fn_execution.aggregate_fast",
+                    sources=["pyflink/fn_execution/aggregate_fast.c"],
                     include_dirs=["pyflink/fn_execution/"]),
                 Extension(
                     name="pyflink.fn_execution.stream",
